+++ R/Utilities.R
@@ -401,16 +401,17 @@ setcolnames_byname <- function(a, colnames){
 #' Rename matrix rows and columns by prefix and suffix
 #' 
 #' It can be convenient to rename rows or columns of matrices 
-#' based on retaining prefixes *before* a separator or suffixes *after* a separator.
+#' based on retaining prefixes *before* a separator or suffixes *after* a separator
+#' in the row and column names.
 #' This functions provides that capability.
 #' 
 #' Note that the `keep` parameter tells which to retain, 
 #' prefixes *before* a separator or suffixes *after* a separator.
 #' 
-#' When `keep` is "prefix", all characters to the right of the first (leftmost) separator and the separator itself are dropped from the names. 
-#' When `keep` is "suffix", all characters to the left of the last (rightmost) separator and the separator itself are dropped from the names. 
+#' When `keep` is "prefix", all characters to the left of the first (leftmost) separator are retained in the row and/or column names.
+#' When `keep` is "suffix", all characters to the right of the last (rightmost) separator are retained in the row and/or column names.
 #' 
-#' If `sep` is not found in a name, that name is unchanged.
+#' If `sep` is not found in a row and/or column name, that name is unchanged.
 #' 
 #' @param a a matrix or list of matrices whose rows or columns will be renamed
 #' @param sep a string that identifies the separator between prefix and suffix
@@ -422,7 +423,13 @@ setcolnames_byname <- function(a, colnames){
 #' @export
 #'
 #' @examples
-rename_to_pref_suff_byname <- function(a, sep, keep, margin = c(1, 2)) {
+#' m <- matrix(c(1, 2, 
+#'               3, 4, 
+#'               5, 6), nrow = 3, byrow = TRUE, 
+#'             dimnames = list(c("a -> b", "r2", "r3"), c("a -> b", "c -> d")))
+#' rename_rowcol_to_pref_suff_byname(m, sep = " -> ", keep = "prefix")
+#' rename_rowcol_to_pref_suff_byname(m, sep = " -> ", keep = "suffix")
+rename_rowcol_to_pref_suff_byname <- function(a, sep, keep, margin = c(1, 2)) {
   
   rename_func <- function(a, sep, keep = c("prefix", "suffix"), margin = c(1, 2)) {
     # At this point, a should be a single matrix.