+++ NEWS.md
@@ -1,7 +1,7 @@
 * Added additional tests for bug fixes and new features.
     * Now up to 782 tests, all passing.
     * Code coverage remains at 100%.
-* Added new function `rename_to_pref_suff_byname()` which 
+* Added new function `rename_rowcol_to_pref_suff_byname()` which 
   renames rows or columns to prefixes or suffixes in row and column names.
 * Fixed a bug in `clean_byname()` which caused 
   a `NULL` response when unnamed rows or columns were present.